refactor(ARCHITECT-002): split OrderDetail.jsx into sub-components
Extract 6 sub-components from OrderDetail.jsx (1696 → 924 lines):
- ProductionStatusCards (ProductionProgressSummary + ProductionOrderStatusCard)
- MaterialRequirementsSection (table + shortage indicators)
- CapacityRequirementsSection (routing capacity table)
- PaymentsSection (summary, history, record/refund)
- ShippingAddressSection (display + inline edit with own API call)
- OrderModals (CancelOrderModal + DeleteOrderModal)

Also fixed dynamic Tailwind classes in ProductionOrderStatusCard with
static class maps to avoid production purge issues.
